config/requirements.txt
# 发布配置依赖
pyyaml>=6.0
